Wrapper function for the chi2 function from the discretization package.

This function discretizes a training set using the chi2 method and the user-provided parameters and chooses the best discretization scheme among them based on a user-provided criterion and eventually a test set.

chi2_iter(predictors, labels, test = TRUE, validation = TRUE,
  criterion = "gini", param = list(list(alp = 0.001, del = 0.5)))

Arguments

predictors
The matrix array containing the numeric attributes to discretize.

labels
The actual labels of the provided predictors (0/1).

test
Boolean : True if the algorithm should use predictors to construct a test set on which to search for the best discretization scheme using the provided criterion (default: TRUE).

validation
Boolean : True if the algorithm should use predictors to construct a validation set on which to calculate the provided criterion using the best discretization scheme (chosen thanks to the provided criterion on either the test set (if true) or the training set (otherwise)) (default: TRUE).

criterion
The criterion ('gini','aic','bic') to use to choose the best discretization scheme among the generated ones (default: 'gini'). Nota Bene: it is best to use 'gini' only when test is set to TRUE and 'aic' or 'bic' when it is not. When using 'aic' or 'bic' with a test set, the likelihood is returned as there is no need to penalize for generalization purposes.

param
List providing the parameters to test (see ?discretization::chi2, default=list(list(alp=0.001, del=0.5))).

Examples

# Simulation of a discretized logit model
set.seed(1)
x = matrix(runif(300), nrow = 100, ncol = 3)
cuts = seq(0,1,length.out= 4)
xd = apply(x,2, function(col) as.numeric(cut(col,cuts)))
theta = t(matrix(c(0,0,0,2,2,2,-2,-2,-2),ncol=3,nrow=3))
log_odd = rowSums(t(sapply(seq_along(xd[,1]), function(row_id) sapply(seq_along(xd[row_id,]),
function(element) theta[xd[row_id,element],element]))))
y = stats::rbinom(100,1,1/(1+exp(-log_odd)))

chi2_iter(x,y)

#> Warning: Speedglm not installed, using glm instead (slower).

#> 
#> Call:  stats::glm(formula = labels ~ ., family = stats::binomial(link = "logit"), 
#>     data = Filter(function(x) (length(unique(x)) > 1), as.data.frame(sapply(disc[[i]]$Disc.data, 
#>         as.factor))))
#> 
#> Coefficients:
#> (Intercept)          V12          V22          V32  
#>        2.63        -3.46       -20.47        -2.44  
#> 
#> Degrees of Freedom: 59 Total (i.e. Null);  56 Residual
#> Null Deviance:	    83.2 
#> Residual Deviance: 32.9 	AIC: 40.9